Change map size_type to uint32

+++ engine/include/RedEngine/Core/Container/Map.hpp
@@ -49,7 +49,7 @@ public:
         ValueT second;
     };
 
-    using size_type = uint64;
+    using size_type = uint32;
     using value_type = KeyValuePair;
     using ThisType = Map<KeyT, ValueT, HashOpT, EqualsOpT>;
